В методах Use() тип возврата заменен на WebApplication
* тип возврата заменен на WebApplication
* класс расширений CORS разделен на конфигурируемые и разрешительные методы
* изменения в xml-комментариях

https://github.com/andrey-aka-skif/AndreyAkaSkif.ServiceDefaults/issues/17
